+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 1.35
+* mobx5 support
+* Introduced observable `type` field prop
+* Removed `label` fallback to field `name`
+* fix: #432 #415 #355 #347 #425 #429 #416
+
 # 1.34
 * Added `struct` constructor prop
 * Reimplemented add() method